refactor(gas-service): simplify API
The GasService was initially designed such that multiple instances could
coexist. This was due to the belief that we'd have multiple
solana-axelar relayers, each with it's own GasService state. After the
introduction of the Operators contract, it was clear that that belief
was wrong. Thus, we can simplify some APIs by removing the need for the
user to provide the program id and config PDA, since these can be
derived easily internally.

Signed-off-by: Guilherme Felipe da Silva <gfsilva.eng@gmail.com>